Make fluid ounce usage consistent to US fl. oz.
This fix changes all uses of fluid ounces to be US fl. oz., whereas
before it was a mix of Imperial and US. It also introduces automatic
millilitre calculation, and adds the litre unit as an option.

Provide the option to display just oz, just mL, or both, with automatic
conversion to L for sizes above 1000 mL inclusive.

Fixes issue #6.